+++ psecagg/src/main/java/edu/bjut/psecagg/messages/MsgResponseRound2.java
@@ -3,17 +3,17 @@ package edu.bjut.psecagg.messages;
 import java.util.ArrayList;
 
 public class MsgResponseRound2 {
-    private ArrayList<Long> u3ids;
+    private ArrayList<Integer> u3ids;
 
-    public MsgResponseRound2(ArrayList<Long> u3ids) {
+    public MsgResponseRound2(ArrayList<Integer> u3ids) {
         this.u3ids = u3ids;
     }
 
-    public ArrayList<Long> getU3ids() {
+    public ArrayList<Integer> getU3ids() {
         return u3ids;
     }
 
-    public void setU3ids(ArrayList<Long> u3ids) {
+    public void setU3ids(ArrayList<Integer> u3ids) {
         this.u3ids = u3ids;
     }
 }
