+++ psecagg/src/main/java/edu/bjut/psecagg/messages/MsgRound0.java
@@ -4,23 +4,23 @@ import it.unisa.dia.gas.jpbc.Element;
 
 public class MsgRound0 {
 
-    private long id;
+    private int id;
     private Element cPk_u;
     private Element sPk_u;
     private Element sigma_u;
 
-    public MsgRound0(long id, Element cPk_u, Element sPk_u, Element sigma_u) {
+    public MsgRound0(int id, Element cPk_u, Element sPk_u, Element sigma_u) {
         this.setId(id);
         this.setcPk_u(cPk_u);
         this.setsPk_u(sPk_u);
         this.setSigma_u(sigma_u);
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
